+++ docs/site/content/en/docs/overview/changelog.md
@@ -23,6 +23,7 @@ August 2023
 - Added [set presence endpoint](https://waha.devlike.pro/docs/how-to/presence)!
   - With `online` presence you can show your status as online [#121](https://github.com/devlikeapro/whatsapp-http-api/issues/121)
   - With `offline` presence you can get notifications for your device [#28](https://github.com/devlikeapro/whatsapp-http-api/issues/28)
+- Added security settings for edit group info `PUT /api/{session}/groups/{groupId}/settings/security/info-admin-only` - updates the group settings to only allow admins to edit group info (title, description, photo).
 
 ## 2023.7
 July 2023